+++ CBA.Core/Implementation/LoanAccountConfig.cs
@@ -11,7 +11,7 @@ namespace CBA.Core.Implementation
     {
         public int Id { get; set; }         
         //[Range(0.1,10)]
-        public int DInterestRate  { get; set; }
+        public decimal DInterestRate  { get; set; }
 
         public GlAccount InterestIncomeGl { get; set; }
